testdrive: add support for retry mofidiers
This PR is in preparation for the changes in tests that will come with
the correctness proprety PR, namely that the vast majority of testdrive
commands won't have to be retried to get the expected result.

In this PR I'm adding support for an additional modifier (similar to the
`version<=123` one) that can set whether a command should be retried or
not.

Here are some examples of what they will look like:

```
>[retry] SELECT 1
1

![retry, version>=5900] SELECT 1
contains: foobar
```

I've set the default value to be true in this PR since we don't have the
correctness property work done yet but I will flip that to false
together with an adjustment of any affected test in that PR.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>