adapter: Include relations in timedomain
Previously, a relation was excluded from a transactions timedomain
if the relation was indexed. As a result if the index was dropped in
the middle of a transaction then the relation was un-queryable in the
transaction.

This commit includes all storage collection backed relations in the
timedomain, even if there's an index.

Works towards resolving #21844