kafka: respect low watermark when setting up capabilities
It can happen that a source is pointed to a topic that used to have
messages (i.e high watermark is non-zero) but due to compaction all
messages are deleted (i.e its low watermark equals the high watermark).

The code on main unconditionally started ingesting a partition at offset
0, expecting to eventually reach the high watermark. If the topic was
completely empty due to compaction this never happened, causing the
source to permanently be unqueryable (or until a single message got
published to the empty partition).

This PR fixes this issue by respecting the low watermark when a new
partition is discovered, fast forwarding the data capability to the
offset of the low watermark instead of zero.

Fixes #21860

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>