DataflowError decoding fallback to the legacy format
In #21510 we introduced a change to the `DataflowError` protos that
wasn't backward compatible. This made Mz panic on some persisted
`DataflowError`s because it was unable to decode them anymore.

This commit fixes the `DataflowError` (and `SourceData`) decoding by
falling back to `*Legacy` proto versions in case decoding with the new
versions fails.