sql: Execute entire transaction on single cluster (#21854)
This commit updates the transaction logic so the entire transaction is
executed on a single cluster. Users cannot change clusters
mid-transaction and auto-routing is disabled mid-transaction.

The first query in a transaction determines the time domain of the
transaction. The timedomain only contains indexes on the active cluster
of the first query. Switching clusters will almost always lead to
timedomain errors. Additionally, some serving layer work will be moved
to clusters in the future. Switching cluster mid-transactions will
complicate that work.

Including multiple cluster indexes in a single transaction's timedomain
has negative use-case isolation implications because we need to take
read holds on all indexes.

Resolves #21839