ci: require backward compatibility for all protobufs
We recently had an incident caused by introducing a backward
incompatible change to a protobuf definition. So far we've been
differentiating between ptotobufs that are persisted from protobufs that
don't. We've been liberally introducing backwards incompatible changes
to the ones that are not persisted and been careful with the ones that
are.

Unfortunately due to a combination of human error and a misconfiguration
of the backward compatibility checker we shipped backward incompatible
change.

This PR makes it so backward compatibility is required for all protobuf
definitions for two reasons:

1. It will start a culture shift of everyone getting into the habit of
   making backward compatible changes.
2. It will make it much harder for the CI check to fail as there will be
   no exceptions.

There is the possibility that we truly want to ship a backward
incompatible change between for non-persisted protobufs in the future
(e.g for the controller <-> clusterd protocol) and we might need to
relax this policy. If and when this happens we can reconsider our tool
configuration and hopefully by that point some muscle memory will have
already been infused to the team.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>