Introduced fixed-length representations
This commit introduces fixed-length representations in mz_repr. These
representations have the properties that: 1. The data stored in them
includes no tags; 2. They can be converted to Row by supplying a
schema; 3. They can be created from a Row provided a validating schema.
The first implementation provided is for an N-byte representation, in
which N is a compile-time parameter. This implementation accepts
single-column schemas wherein the type with nullability can be stored
in N bytes. The intended first use is with a 9-byte representation, so
a typedef with N=9 (Bytes9) is provided.