catalog: Remove old role attribute comment (#21920)
This commit removes an old TODO comment about converting RoleAttributes
to a bitflag. RoleAttributes used to be a struct containing many
different bool fields, and would have benefited from being a single u64
bitflag. Currently, it's only a single boolean field which wouldn't
benefit from being converted to a bitflag. It's not clear that all
fields added in the future will aso be boolean. Therefore, converting it
to a bitflag would be a premature optimization.