stash: split test_stash into multiple tests
These can now be run in parallel.

Improve DebugStashFactory to allow it to make all stash variants.

Additionally (and sadly missed missed until now): teach the Consolidator
about schemas. It is fairly scary we have not done this until now, which
is likely a result of passing schemas in URLs. Probably only tests ever
needed this. Now that we use the DebugStashFactory for things which
creates a random schema each time, it was required to get some tests to
pass.

See #21891