explain: add `ExprHumanizerExt` helper
When explaining `CREATE <item>` statements, the `drain_all` method
will need to use an `ExprHumanizer` instance that reports the explained
item as created even if the item is not present in the the backing
catalog state.

This commit introduces an `ExprHumanizerExt` struct that can be used to
report non-existing items (passed as a `BTree<GlobalId, TransientItem>`)
as present to `ExprHumanizer` clients.