Specialize 9-byte key arrangements
This commit enables specialization of arrangements for 9-byte keys
representable by the corresponding fixed-length type. The commit is
breaking in the sense that for joins to work properly, all arrangements
need to be specialized to such keys in the same way. However,
arrangements coming out of reductions are still not specialized, which
will lead to a panic upon execution.