postgres-client: add PostgresClient in new mz-postgres-client
This extracts the re-usable parts out of persist's `PostgresConsensus`,
such that we can re-use it in the Postgres/CRDB-backed `TimestampOracle`.

Right now, `PostgresClient` is a leaky abstraction: the only interface
is `get_connection()`, which directly exposes the deadpool API and
persist will still need all the code for mapping deadpool/postgres
errors to `ExternalError`. In its current state, this largely exists for
code sharing.

We could, as a next step hide deadpool and only expose methods for
running things like `execute()`, or `query()`. Or we could rename
`PostgresClient` to `PostgresPool` and have `get_connection()` return a
new `PostgresClient` that exposes only those methods that we need _and_
does the mapping between deadpool/postgres errors and `PostgresError`.
Thus completely hiding deadpool/postgres, including its errors from the
user.

We will change `PostgresConsensus` to use this in the next commit.