test: (Hacky) fix for slow test cleanups (#22025)
Running tests in parallel has been causing some deadlock/starvation
issue that we haven't been able to figure out. For some reason,
uncommitted transactions are being left open which causes the schema
cleanup in the Drop impl of DebugStashFactory to block for an enormous
amount of time (many minutes). Setting a small idle transaction timeout
globally seems to resolve the issue somehow. This is a gross hack and
we don't really understand what's going on, but for now it resolves
issues in CI while we debug further.

Works towards resolving #21891