persist-txn: initial perf tuning
Add a cli tool for easy iteration on high-level perf, including the
ability to send traces to Tempo.

Fix a few instances of things that run serially instead of concurrently.
Also reduce CaS-mismatch retries by starting with an expected upper that
is more likely to be correct (i.e. the one known by pubsub).