persist: add a WriteHandle::wait_for_upper_past method
Almost entirely code movement of `next_listen_batch` from `WriteHandle`
to `Machine`.

This is used by persist-txn to wait for some write to be applied. We
could _almost_ use ReadHandle::snapshot for this (by not fetching the
data and immediately returning the leases), except that it requires the
shard since to be less_equal the as_of. This fails in the following
scenario:

- We're trying to serve a data shard snapshot at ts 9.
- The most recent write before that point was at 5.
- The data shard since is at 7 (allowed by persist-txn).
- We need to wait for the write at 5 to be applied, which is done by
  calling this new method with 5.
- Then, because persist-txn knows there are no writes in `(5, 9]`, it
  does an empty CaA of `[6, 10)`.
- Now persist-txn can proceed to take a snapshot at 9 as normal.