zippy: Exercise cluster unification
Randomly create sources either in the dedicated `storage` cluster,
as would be the case with most customers, or in the `default` cluster,
where they will co-mingle with the materialized views and thus
exercise cluster unification.

Zippy scenarios that create additional replicas to the `defaut` cluster
will continue sending all the sources to the `storage` cluster exclusively.

Relates to #21846