Feature flag for jemalloc profiling
This commit adds a feature flag `enable_jemalloc_profiling` to control
whether jemalloc profiling is enabled. The intention is to remove this
flag again once we are convinced that jemalloc profiling doesn't cause
performance regressions.

The flag needs to be applied to all the `environmentd` and `clusterd`
processes. The coordinator applies it to `environmentd` directly and
sends a config update to the compute replicas to apply the change to
`clusterd`. No config update is sent to storage replicas, because it is
not necessary: Each `clusterd` has both a compute and a storage replica
running and only one of them needs to apply profiling config changes.