catalog: Add durable catalog snapshot type (#21922)
This commit adds a durable catalog snapshot type that represents a
complete snapshot of the on-disk state. A trait method was added to
ReadOnlyDurableCatalogState to retrieve the snapshot and the
transaction constructor method was updated to be built from a snapshot.

This achieves a couple of useful things:
- Organizes the code a little bit.
- Puts in some of the foundations for a DurableCatalogState
implementation to maintain an in-memory cache.
- Puts in some of the foundations to allow a transaction to be built
from memory, without going to a durable store.