persist: take a PersistClient future in shard_source
Previously, it required an `Arc<PersistClientCache>` and a
`PersistLocation`. This was mainly becuase that's what persist_source
takes, which in turn is because at dataflow rendering time, it's
difficult to evaluate Futures (e.g. `cache.open(location).await`).

Now, instead take any future that gives you a PersistClient (technically
a Fn that returns these futures because we need two of them). This will
make it easier to use `shard_source` in tests as I'm about to do in an
upcoming persist-txn PR.