timely-util: Wakers should anticipate hung up receivers
We don't have any guarantees about how long the `Waker` will be held for
and so we must be prepared for the receiving end to have hung up when we
finally do get woken. This can happen if by the time the waker is called
the receiving timely worker has been shutdown. For this reason we ignore
the activation error.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>