adapter: optimize views in background during bootstrap
Change catalog opening to optimize views in background threads. This
allows the various open work to proceed in parallel to optimization,
which can take many seconds in debug.

The actual method this is accomplished is unpleasantly ugly. It is done
by accumulating JoinHandles during open and join'ing them before open
returns. The code assumes (and I verified) that nothing depends on the
optimized view exprs during open, so it's safe to join at the end of
that function. There's no code guarantee this remains the case, and race
conditions could easily make this appear to work until it blows up
somewhere. Plumbing the JoinHandle up from parse_item also feels overly
hacky.

One alternative was to do just-in-time optimizing during first use. But
because the optimization can return an error, this is not a great
approach because most of the code paths that need the optimized
expression don't return errors. Also it'd be very weird to return an
error after successfully creating an object if using it would always
produce an error.

On my machine, debug envd startup went from 5.5s to 2.0s. This should
speed up tests and our local dev cycle for essentially zero production
performance cost (and will speed up envd boot by just a bit there too).