prof: require activated heap profiling
This commit makes all HTTP endpoints that allow dumping a heap profile
check whether heap profiling is activated before doing so.

The main intent is to be able to gate the `mz-build-id` invocation
behind the `enable_jemalloc_profiling` feature flag, to make it possible
to opt out of the execution of the unsafe code contained therein. But it
generally seems like a good idea to return an error rather than a
useless heap profile.