adapter: fork old implementation for `CREATE MATERIALIZED VIEW`
Fork the old implementations of

- `Coordinator::bootstrap` (the `CatalogItem::MaterializedView` part),
- `Coordinator::sequence_create_materialized_view`, and
- `Coordinator::explain_create_materialized_view_optimizer_pipeline`,

by duplicating code and dispatching between the old and the new code
path based on the value of the `enable_unified_optimizer_api` feature
flag. At the moment the two code paths are still identical, but this
will change with the next commit.

See the rollout proposal in #20569 for details.