Introduce flag enable_specialized_arrangements
This commit introduces the feature flag
enable_specialized_arrangements, which gates the functionality of type
specialization for arrangements. The commit restores full functionality
by making all arrangements employ Row key and values as previously. At
the same time, it enables incremental development of the type
specialization feature for arrangements. At the current stage, enabling
type specialization for arrangements will lead to panics in a
non-trivial subset of queries. To document the subset of functionality
that presently works with arrangement specialization, this commit also
introduces a testdrive test.