Ensure completeness of mz_dataflow_arrangement_sizes
This commit adjusts the definition of the
`mz_dataflow_arrangement_sizes` view to include all active dataflows,
not only the ones referenced from the catalog. This is more in line with
the implicit expectation users will have from this view, based on its
name and description in the docs.

Previously, `mz_dataflow_arrangement_sizes` included a JOIN with
`mz_objects` which caused all dataflows not contained in the latter to
be excluded. This lead to some confusion when people where assuming high
memory usage wasn't caused by arrangements because they didn't show up
in `mz_dataflow_arrangement_sizes`.

This commit adapts the view definition to not join against `mz_objects`
anymore and to produce the dataflow name instead of the object name.
We could alternatively use a LEFT JOIN against `mz_objects` and set
missing object names to NULL. We prefer using dataflow names because:

 * Dataflow names are always present and not less readable than catalog
   object names (e.g. 'Dataflow: materialize.public.my_idx' instead of
   'my_idx').
 * In case of dataflows that belong to dropped objects or objects that
   were never written to the catalog (SELECTs, SUBSCRIBEs) they still
   provide useful information (e.g. 'Dataflow: subscribe-t1' instead of
   NULL).
 * Having a dataflow name in a view called
   `mz_dataflow_arrangement_sizes` is arguably less surprising than
   having a catalog object name.
 * The view definition is simplified.