storage: heal shards that contain backwards incompatible errors
This PR implements the remediation of having persisted backward
incompatible protobuf data in certain errored upsert sources. I will
explain the general strategy. The protobuf message that contained the
backward incompatible field was `ProtoUpsertError` and specifically its
`value` field that changed from being a `ProtoDataflowError` in `0.68.1`
to being a `ProtoDecodeError` in `0.69.1`.

This PR is changing the `value` field to be of type
`mz_storage_types.shim.ProtoUpsertValueErrorShim`. This type lives in
its own file (`shim.proto`) which is necessary for `prost` to correctly
generate the code but the generated `shim.rs` file is *not* included in
the module tree with the `include!()` macro that we do normally. This
allows us to autogenerate all the protobuf types with one hole that we
need to fill in.

This PR includes a manually created `shim.rs` file that manually
declares the `ProtoUpsertValueErrorShim` type to be an enum that has two
variants, either `New` or `Legacy`. The `New` variant contains the real
proto type that `0.69.1` uses and the `Legacy` variant contains the real
proto type that `0.68.1` used (renamed to `ProtoUpsertValueErrorLegacy`
in this PR).

This manual declaration also includes a manual `prost::Message`
implementation which is what `prost` autogenerates for you from all the
proto files. This `Message` implementation mostly delegates to the inner
proto types by matching on the variant and forwarding the call. There is
only one call that we override and that is `merge_field`.

`Message::merge_field` is called a protobuf message is being
deserialized and its job is to mutate `self` and add information for the
field in question. The `value` field of `ProtoUpsertError` (careful,
this is NOT the same as `ProtoUpsertValueError`) has the tag number `1`
for both new and legacy variants. Therefore, when `merge_fields` is
called with `tag == 1` it's when we deploy a bunch of heuristics and
application specific knowledge to understand if the bytes presented to
us are of the new or the old variant.

If you're following along we've now have created a special `Proto`
struct that can understand what type the bytes are presented to it and
neatly encodes this information as an enum.

Moving along, we need to look at `UpsertvalueError` the "real" rust
struct that we work with in memory. This struct gained a new field
called `is_legacy_dont_touch_it` which is set by its
`RustType::from_proto` implementation. This is simply turning the `enum`
information that existed in `ProtoUpsertValueErrorShim` into a boolean
field. The reason for not making `UpsertValueError` also an enum is due
to the number of code changes that would have to go with it.

Ok, so now we have an in-memory `UpsertValueError` type that can
remember if it was created by a legacy encoding. The final piece of the
recipe is in the upsert operator. When the upsert operator starts up and
rehydrates itself from a snapshot it takes note of all the legacy errors
that it got and their associated diffs. When the snapshot phase is over,
the operator will emit retractions (i.e negated diffs) for all observed
legacy errors and insertions for non-legacy ones. This means that
shortly after that moment we will permanently record in persist a
retraction of the old incompatible bytes and an insertion of the new,
correct bytes.

Finally, when compaction eventually catches up and goes over these diffs
it will completely eliminate the legacy encoding and at that point we
should be able to delete the monstrosity presented here.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>