storage: reclock snapshot to minimum timestamp
This fixes correctness property #2, albeit a bit indirectly, by
reclocking the initial snapshot of each source to the minimum timestamp.
This approach sidesteps the need for intializing the since frontier of a
source in coordination with its remap trace and it additionally enables
users to use temporal filters to place each of the ingested records to
the appropriate time.

Closes #13639
Closes #11632

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>