catalog: Add persist catalog command line args
This commit adds command line arguments to environmentd, so that we can
determine the implementation of the `DurableCatalogState` at runtime.
This required switching from generic methods to dynamic dispatch with
Boxes in various places.

Additionally, this commit switches from using the term "environment_id"
to "organization_id" when referencing an organizations UUID because it
is more accurate.

Works towards resolving #22392