test: Speed up auth tests (#22150)
The auth tests sometimes flake due to an expired authentication.

The auth tests start a fake materialize server that responds to Frontegg
requests with generated claims. The tests use a constant,
EXPIRES_IN_SECS, that is passed into the fake test server, so that all
claims are returned with an exp field set to the current time minus
EXPIRES_IN_SECS. The exp field is The time at which the claims expire,
represented in seconds since the Unix epoch. Essentially EXPIRES_IN_SECS
represents how many seconds the claims are valid as of when they're
created.

The tests also use a constant, REFRESH_BEFORE_SECS, that determines how
many seconds before the claims expire to start refreshing the claims.
Essentially, we attempt to refresh a token REFRESH_BEFORE_SECS seconds
before it expires.

The test flake happens when the refresh takes long enough that by the
time the refresh is complete the token has already expired. As a result,
as long as REFRESH_BEFORE_SECS <= EXPIRES_IN_SECS, the flakiness is
completely determined by REFRESH_BEFORE_SECS, since that is the amount
of time we have to complete the refresh.

This commit lowers EXPIRES_IN_SECS from 40 to 20, and 
REFRESH_BEFORE_SECS from 20 to 15 to decrease the time we wait for a
token refresh without affecting test flakiness.

Works towards resolving #22145