compute: add timely_step_duration_seconds histogram
In incident 70, we found that timely driven Futures were not promptly
being polled after being woken. This was presumably because some
operator was not regularly yielding. Not yielding also causes issues in
query interactivity (i.e. because peeks are included in this). The
introspection sources have full-fidelity data around operator run times,
but we duplicate a proxy for them into prometheus so we can alert on it
and easily scan it in investigations.

Concretely, a new histogram is added for step_or_park durations in
compute timely clusters. This doesn't tell us anything about individual
operators (and including the park time is undesirable) but it has the
benefit of being very easy yet still useful. In particular, anything
over a second or two is probably a hint to investigate.