stash: split out a stash-types crate
This allows for faster iteration on mz-stash itself, as is happening in
Pv2.

Conceptually, stash is only used in envd (by the catalog and by the
storage controller) but it was being pulled in by all sorts of crates
for the objects protos. Separate out the latter (plus StashError) into a
new mz-stash-types crate that is still depended on by lots of things,
but changes less often.

The various old versions of objects.proto want to be next to it, so pull
them into mz-stash-types, too. In an ideal world, we'd probably move the
vX_to_vX+1 upgrades also, but they depend on actual stash things (e.g.
Transaction), so leave them where they are.

Finally, any given metric can only be registered once (otherwise we get
a panic), but we need two stashes, so also pull out the stash Metrics
into mz-stash-types. Ideally, this would have stayed where it was. A
possible alternative to this would be something like putting a type-safe
cache of registered metrics objects on MetricsRegistry.

Timing on my M1 laptop of bin/environmentd after invalidating stash went
from 2m12s to 53s. A large part of this is likely that we've entirely
eliminated clusterd, including linking.

```
touch src/stash/src/lib.rs && time ./bin/environmentd -- --nope
```