RBAC: Always enable RBAC for system roles
This commit ensures that RBAC is always enabled for system roles. This
allows us to ensure that we always gate the privilege of system roles,
even when RBAC is disabled, without needing specialized hacky privilege
functions. The specialized hacky privilege functions are not as
expressive as RBAC and don't always provide us with the proper
semantics.

Fixes #22245