Enable introspection debugging by default on new replicas
This changes the behavior when creating new replicas: Instead of defaulting
to false for introspection debugging, we default to true. This causes
the introspection dataflows show up in introspection sources, which can
cause more work, but also allows to attribute processing time where we
currently tap in the dark.

This PR can cause a higher CPU utilization, and quick manual tests showed
an increase in the 3-5% range, although this needs to be validated again.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>