ASSERT NOT NULL in MVs (#22248)
This introduces the `ASSERT NOT NULL` option on materialized views. It
is intended to allow users to force elements in the schema of
corresponding sinks to be non-nullable.

We also check at runtime that all the values are actually non-null, and
emit an error otherwise. This is done with a bit of code in the compute
sink, which feels like an abstraction violation: ideally, I would have
liked to make this a separate `Assert` MIR node, and had rendering not
know anything about it. But it's currently considered impossible to make
any guarantees about nullability inference on the MIR program, so I
decided against trying to implement this in that way.

Note that this still doesn't fully solve the problem of nullability
control in sinks, as we lack a good way to subscribe to just the errors
in the MV.

### Motivation

  * This PR fixes a recognized bug.
https://github.com/MaterializeInc/materialize/issues/21722
 
### Checklist

- [x] This PR has adequate test coverage / QA involvement has been duly
considered.
- [x] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [x] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [x] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - Add `ASSERT NOT NULL` option for sinks